Added controller logic and test framework.

+++ src/main/java/edu/mayo/cts2/framework/plugin/namespace/service/impl/NamespaceMaintenanceServiceImpl.java
@@ -1,13 +1,15 @@
 package edu.mayo.cts2.framework.plugin.namespace.service.impl;
 
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.Assert;
 
 import edu.mayo.cts2.framework.model.service.namespace.MultiNameNamespaceReference;
 import edu.mayo.cts2.framework.plugin.namespace.dao.NamespaceRepository;
 import edu.mayo.cts2.framework.plugin.namespace.model.Namespace;
-import edu.mayo.cts2.framework.plugin.namespace.service.NamespaceMaintenanceService;
+import edu.mayo.cts2.framework.service.namespace.NamespaceMaintenanceService;
 
 @Component
 public class NamespaceMaintenanceServiceImpl implements NamespaceMaintenanceService {
@@ -22,7 +24,52 @@ public class NamespaceMaintenanceServiceImpl implements NamespaceMaintenanceServ
 
 	@Transactional(readOnly = false)
 	public void delete(String uri) {
-		this.namespaceRepository.delete(uri);
+		if(this.namespaceRepository.exists(uri)){
+			this.namespaceRepository.delete(uri);
+		}
 	}
 
+	@Transactional(readOnly = false)
+	public void addLocalName(String uri, String localName, boolean setPreferred) {
+		Assert.hasText(localName, "'LocalName' must not be empty.");
+		
+		Namespace namespace = this.namespaceRepository.findOne(uri);
+		if(namespace == null){
+			namespace = new Namespace(uri);
+			namespace.setPreferredName(localName);
+			
+			this.namespaceRepository.save(namespace);
+		} else {
+			//check if is already there
+			if(! StringUtils.equals(namespace.getPreferredName(), localName)){
+				
+				//check if it is an alternate name
+				if(namespace.getAlternateNames().contains(localName)){
+					if(setPreferred){
+						namespace.getAlternateNames().remove(localName);
+						namespace.setPreferredName(localName);
+						
+						this.namespaceRepository.save(namespace);
+					}	
+				} else {
+					if(setPreferred){
+						String currentPreferred = namespace.getPreferredName();
+						if(StringUtils.isNotBlank(currentPreferred)){
+							namespace.getAlternateNames().add(currentPreferred);
+						}
+						namespace.setPreferredName(localName);
+					} else {
+						String currentPreferred = namespace.getPreferredName();
+						if(StringUtils.isBlank(currentPreferred)){
+							namespace.setPreferredName(localName);
+						} else {
+							namespace.getAlternateNames().add(localName);
+						}
+					}
+					
+					this.namespaceRepository.save(namespace);
+				}
+			}
+		}
+	}
 }
